UPD. Chief page w/ router 6 + minor adjustments
