+++ src/utils/Menu.js
@@ -2,13 +2,11 @@ import {useState} from 'react';
 
 import {Drawer, IconButton, Box, List, ListItem, ListItemText} from '@mui/material';
 import {ArrowBack as ArrowBackIcon, Menu as MenuIcon,} from '@mui/icons-material';
- import {Link} from "react-router-dom";
 
 
 export default function Menu({menuItem}) {
-    // const { path, url } = useRoutes();
-    const [state, setState] = useState(false);
 
+    const [state, setState] = useState(false);
     const toggleDrawer = (foo) => (event) => {
         if (
             event.type === 'keydown' &&
@@ -22,21 +20,19 @@ export default function Menu({menuItem}) {
     return (
         <>
             <IconButton onClick={toggleDrawer()}>
-                {state ? <ArrowBackIcon/> : <MenuIcon/>}
+                {state ? <ArrowBackIcon fontSize={'large'}/> : <MenuIcon fontSize={'large'}/>}
             </IconButton>
             <Drawer open={state} onClose={toggleDrawer()}>
-                {/*Todo: redo with List and ListItem*/}
                 <Box sx={{width: 250, my: 5}}>
                     <List>
                         {menuItem.map((anchor, index) => (
-
-                            <ListItem button={true} sx={{mx: 2}}
-                                      onClick={toggleDrawer(anchor.func)} key={index}
-                                       component={Link} to={`${anchor.href}`}
+                            <ListItem
+                                button={true}
+                                sx={{mx: 2}}
+                                onClick={toggleDrawer(anchor.func)} key={index}
                             >
                                 <ListItemText primary={anchor.name}/>
                             </ListItem>
-                            // </Link>
                         ))}
                     </List>
                 </Box>
